move to subpackage and added reflection proxy benchmark
